Reimplement stack downstacking as a :_=<sha> filter op
Move the downstack rebase logic out of josh-changes into josh-core's filter
pipeline as a new Op::Downstack, exposed in flang as ":_=<sha>". This lets
the operation participate in the filter cache via the transaction, and lets
the trailer-parsing helpers move alongside it into josh-core::trailers.

Change: downstack-as-filter-op
