+++ josh-proxy/src/upstream.rs
@@ -495,6 +495,7 @@ pub fn process_repo_update(repo_update: RepoUpdate) -> anyhow::Result<String> {
 
         let to_push = build_to_push(
             transaction.repo(),
+            &transaction,
             &push_mode,
             &baseref,
             &ref_with_options,
